requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
mysqlclient==2.1.0
PyJWT==2.4.0